refactor(all): seperate browserFeature from browser for test
Ref #210
