+++ src/YawPitchControl/YawPitchControl.js
@@ -1,6 +1,6 @@
 import Component from "@egjs/component";
 import Axes, {PinchInput, MoveKeyInput} from "@egjs/axes";
-import {getComputedStyle, SUPPORT_TOUCH, SUPPORT_DEVICEMOTION} from "../utils/browser";
+import {getComputedStyle, SUPPORT_TOUCH, SUPPORT_DEVICEMOTION} from "../utils/browserFeature";
 import WheelInput from "./input/WheelInput";
 import TiltMotionInput from "./input/TiltMotionInput";
 import RotationPanInput from "./input/RotationPanInput";
