+++ test/unit/SpinViewer/SpriteImage.spec.js
@@ -1,5 +1,5 @@
 import SpriteImage from "../../../src/SpinViewer/SpriteImage";
-import {TRANSFORM, SUPPORT_WILLCHANGE} from "../../../src/utils/browser";
+import {TRANSFORM, SUPPORT_WILLCHANGE} from "../../../src/utils/browserFeature";
 
 describe("SpriteImage", function() {
 	describe("create", function() {
